# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: StellarRoute API
  version: 0.1.0
  description: |
    REST API for DEX aggregation on the Stellar Network.

    StellarRoute aggregates liquidity from Stellar's SDEX (Stellar Distributed
    Exchange) and Soroban AMM pools to find the best available price for any
    token pair.

    ## Asset identifiers

    Assets are encoded in path segments using one of three formats:

    | Format | Example | Meaning |
    |--------|---------|---------|
    | `native` | `native` | XLM (Stellar's native asset) |
    | `CODE` | `USDC` | Issued asset — code only (omits issuer) |
    | `CODE:ISSUER` | `USDC:GA5ZSEJ...` | Issued asset — fully qualified |

    ## Error handling

    All errors return a JSON body with `error` (machine-readable code) and
    `message` (human-readable description).  See the `ErrorResponse` schema
    and the **Error codes** section below for the full list.

    ## Rate limiting

    Endpoints are rate-limited per IP address.  When the limit is exceeded the
    API returns **429 Too Many Requests** with `X-RateLimit-*` headers.

    ## Request correlation

    Clients may optionally send `X-Request-ID` with any request. The API will
    reuse that value, or generate a UUID when the header is omitted, and echo
    the final request ID back in the `X-Request-ID` response header.

    ## Versioning and deprecation

    `/api/v1/*` is in deprecation mode. Responses from v1 routes include
...
    notice period and migration guidance.
...
servers:
...
paths:
...
  /api/v1/pairs:
    get:
      operationId: list_pairs
      summary: List trading pairs
      description: |
        Returns active trading pairs that currently have offers in the SDEX
        orderbook.

        Stable sort order:
        1. `offer_count DESC`
        2. `last_updated DESC`
        3. Canonical asset fields ascending as deterministic tie-breakers

        Pagination:
        - Default page size: `25`
        - Max page size: `100`
        - `cursor` (preferred) or `offset` (fallback)
      tags:
        - trading
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
        - $ref: "#/components/parameters/PageOffset"
      responses:
        "200":
          description: List of active trading pairs
          headers:
            Deprecation:
              $ref: "#/components/headers/DeprecationHeader"
            Sunset:
              $ref: "#/components/headers/SunsetHeader"
            Link:
              $ref: "#/components/headers/DeprecationLinkHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairsResponse"
              example:
                pairs:
                  - base: XLM
                    counter: USDC
                    base_asset: native
                    counter_asset: "USDC:GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN"
                    offer_count: 42
                    last_updated: "2026-02-23T11:59:00Z"
                  - base: USDC
                    counter: BTC
                    base_asset: "USDC:GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN"
                    counter_asset: "BTC:GAUTUYY2THLF7SGITDFMXJVYH3LHDSMGEAKSBU267M2K7A3W543CKUEF"
                    offer_count: 17
                    last_updated: "2026-02-23T11:58:30Z"
                total: 2
                limit: 25
                next_cursor: "25"
                prev_cursor: null
        "400":
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: validation_error
                message: Invalid cursor; expected a numeric offset
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: internal_error
                message: Failed to fetch trading pairs
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/markets:
    get:
      operationId: list_markets
      summary: List markets
      description: Alias of `/api/v1/pairs` with identical pagination and sorting.
      tags:
        - trading
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
        - $ref: "#/components/parameters/PageOffset"
      responses:
        "200":
          description: List of active markets
          headers:
            Deprecation:
              $ref: "#/components/headers/DeprecationHeader"
            Sunset:
              $ref: "#/components/headers/SunsetHeader"
            Link:
              $ref: "#/components/headers/DeprecationLinkHeader"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairsResponse"
        "400":
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/RateLimited"
...
components:
  parameters:
...
    PageLimit:
      name: limit
      in: query
      required: false
      description: Page size. Default 25, max 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    PageCursor:
      name: cursor
      in: query
      required: false
      description: Cursor for the next page (numeric offset token).
      schema:
        type: string
        example: "25"
    PageOffset:
      name: offset
      in: query
      required: false
      description: Offset fallback for pagination when cursor is not used.
      schema:
        type: integer
        minimum: 0
        example: 50
  responses:
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Maximum number of requests allowed in the current window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Number of requests remaining in the current window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Unix timestamp when the rate limit window resets
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: rate_limit_exceeded
            message: Too many requests — please slow down
  headers:
...
    DeprecationHeader:
      description: Indicates that the endpoint version is deprecated.
      schema:
        type: string
        example: "true"
    SunsetHeader:
      description: RFC 7231 IMF-fixdate indicating planned v1 sunset.
      schema:
        type: string
        example: "Wed, 31 Dec 2026 23:59:59 GMT"
    DeprecationLinkHeader:
      description: Link to migration guidance for deprecated v1 routes.
      schema:
        type: string
        example: "</docs/api/v1-migration-guide>; rel=\"deprecation\""
  schemas:
...
    TradingPair:
      type: object
      required: [base, counter, base_asset, counter_asset, offer_count]
      description: A tradeable asset pair with active orderbook depth
      properties:
        base:
          type: string
          description: Human-readable base asset code
          example: XLM
        counter:
          type: string
          description: Human-readable counter asset code
          example: USDC
        base_asset:
          type: string
          description: Canonical base asset identifier (`native` or `CODE:ISSUER`)
          example: native
        counter_asset:
          type: string
          description: Canonical counter asset identifier (`native` or `CODE:ISSUER`)
          example: "USDC:GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN"
        offer_count:
          type: integer
          format: int64
          description: Number of active offers for this pair
          example: 42
        last_updated:
          type: string
          format: date-time
          nullable: true
          description: RFC-3339 timestamp of the most recent offer update
          example: "2026-02-23T11:59:00Z"
    PairsResponse:
      type: object
      required: [pairs, total]
      description: List of active trading pairs
      properties:
        pairs:
          type: array
          items:
            $ref: "#/components/schemas/TradingPair"
        total:
          type: integer
          description: Total number of pairs available across all pages
          example: 2
        limit:
          type: integer
          nullable: true
          description: Effective page size used for this response
          example: 25
        next_cursor:
          type: string
          nullable: true
          description: Cursor token for the next page, if any
          example: "25"
        prev_cursor:
          type: string
          nullable: true
          description: Cursor token for the previous page, if any
          example: "0"
...
    ErrorResponse:
      type: object
      required: [error, message]
      description: Standard API error format
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: validation_error
        message:
          type: string
          description: Human-readable error description
          example: "Amount must be greater than zero"
        details:
          type: object
          nullable: true
          description: Optional structured context about the error
          example:
            parameter: slippage_bps
            min: 0
            max: 10000

  # Authentication is planned for a future release.
  # securitySchemes:
  #   bearerAuth:
  #     type: http
  #     scheme: bearer
  #     bearerFormat: JWT
  #     description: |
  #       API key authentication (Phase 2).
  #       Pass a JWT obtained from the /auth/token endpoint.